pyproject.toml
[project]
name = "corrector"
version = "0.1.0"
description = "Corrector ortográfico/contextual en español con Gemini, entrada DOCX y log de correcciones"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Corrector Team" }]
dependencies = [
  "pydantic>=2.4",
  "python-docx>=1.0.0",
  "google-genai>=0.3.0",
  "python-dotenv>=1.0.1",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.16",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "openai>=1.12.0"
]

[project.scripts]
corrector = "corrector.cli:main"

[tool.setuptools]
packages = ["corrector"]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "black>=24.4.0",
  "ruff>=0.5.5",
  "httpx>=0.27.0",
  "pytest-asyncio>=0.23.7"
]

[tool.black]
line-length = 100
target-version = ["py311", "py310"]
exclude = '''
/(
  \.venv
  | build
  | dist
  | outputs
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["outputs", "build", "dist", ".venv"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "UP", # pyupgrade
]
ignore = [
  "E501",  # handled by black
  "B008",  # allow Depends() in function defaults (FastAPI pattern)
  "E402",  # allow imports after setup code (load_dotenv, logging, sys.path)
]

[tool.ruff.lint.isort]
known-first-party = ["corrector", "server"]
combine-as-imports = true
